fix tele update: ALTER COLUMN user_id SET DATA TYPE BigInt;

+++ aries/modules/sql/afk_sql.py
@@ -2,6 +2,7 @@ import threading
 from datetime import datetime
 
 from sqlalchemy import Boolean, Column, DateTime, Integer, UnicodeText
+from sqlalchemy.sql.sqltypes import BigInteger
 
 from aries.modules.sql import BASE, SESSION
 
@@ -9,7 +10,7 @@ from aries.modules.sql import BASE, SESSION
 class AFK(BASE):
     __tablename__ = "afk_users"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     is_afk = Column(Boolean)
     reason = Column(UnicodeText)
     time = Column(DateTime)
