+++ aries/modules/sql/feds_sql.py
@@ -2,6 +2,8 @@ import ast
 import threading
 
 from sqlalchemy import Boolean, Column, Integer, String, UnicodeText
+from sqlalchemy.sql.sqltypes import BigInteger
+
 from telegram.error import BadRequest, Unauthorized
 
 from aries import dispatcher
@@ -60,7 +62,7 @@ class BansF(BASE):
 
 class FedsUserSettings(BASE):
     __tablename__ = "feds_settings"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     should_report = Column(Boolean, default=True)
 
     def __init__(self, user_id):
