+++ aries/modules/sql/users_sql.py
@@ -2,6 +2,7 @@ import threading
 
 from aries import dispatcher
 from aries.modules.sql import BASE, SESSION
+
 from sqlalchemy import (
     Column,
     ForeignKey,
@@ -11,11 +12,12 @@ from sqlalchemy import (
     UniqueConstraint,
     func,
 )
+from sqlalchemy.sql.sqltypes import BigInteger
 
 
 class Users(BASE):
     __tablename__ = "users"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     username = Column(UnicodeText)
 
     def __init__(self, user_id, username=None):
