+++ aries/modules/sql/warns_sql.py
@@ -1,6 +1,7 @@
 import threading
 
 from sqlalchemy import Boolean, Column, Integer, String, UnicodeText, distinct, func
+from sqlalchemy.sql.sqltypes import BigInteger
 from sqlalchemy.dialects import postgresql
 
 from aries.modules.sql import BASE, SESSION
@@ -9,7 +10,7 @@ from aries.modules.sql import BASE, SESSION
 class Warns(BASE):
     __tablename__ = "warns"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14), primary_key=True)
     num_warns = Column(Integer, default=0)
     reasons = Column(postgresql.ARRAY(UnicodeText))
